+++ src/main/java/main/repository/PostCommentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PostCommentRepository extends CrudRepository<PostComment, Integer> {
+public interface PostCommentRepository extends CrudRepository<PostComment, Long> {
 }
